added confirm deleting

+++ src/main/java/application/graphic/ui/DeleteEntity.java
@@ -4,6 +4,8 @@ import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.util.List;
 
+import javax.swing.JOptionPane;
+
 import application.database.EntityDao;
 import application.interfaces.IEntityFrame;
 
@@ -12,12 +14,20 @@ public class DeleteEntity<T, E extends IEntityFrame<T>> extends ToolButton  {
     super("Удалить", "images/remove.png");
     this.addActionListener(new ActionListener() {
       public void actionPerformed(ActionEvent e) {
-        EntityDao<T> dao = new EntityDao<>(entityClass);
-        List<T> objects = parentWindow.getObjectsToDelete();
-        for(T object: objects) {
-          dao.deleteObject(object);
+        String message = "Вы действительно хотите удалить выбранную(ые) запись(и)?";
+        int result = JOptionPane.showConfirmDialog(null,
+                message,
+                "Подтверждение действия",
+                JOptionPane.YES_NO_OPTION,
+                JOptionPane.QUESTION_MESSAGE);
+        if (result == JOptionPane.YES_OPTION) {
+          EntityDao<T> dao = new EntityDao<>(entityClass);
+          List<T> objects = parentWindow.getObjectsToDelete();
+          for(T object: objects) {
+            dao.deleteObject(object);
+          }
+          parentWindow.setTable();
         }
-        parentWindow.setTable();
       }
     });
   }
